requirements.txt
cssmin==0.2.0
Flask==0.10.1
Flask-Assets==0.10
Flask-Compress==1.1.0
Flask-Script==2.0.5
pyScss==1.3.4
python-decouple==2.3
PyYAML==3.11
